+++ frontend/Dockerfile
@@ -6,7 +6,7 @@ WORKDIR /app/frontend
 COPY package.json .
 COPY yarn.lock .
 
-RUN yarn install 
+RUN yarn install
 
 COPY . .
