+++ frontend/src/Components/Register/type.tsx
@@ -1,0 +1,14 @@
+export type FormInput = { name: string; value: string; placeholder: string; handleChange: (event: ChangeEvent<HTMLInputElement>) => void };
+export type AppAction = {
+  type: string;
+  name: string;
+  value: string;
+};
+export interface ReducerType {
+  username: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  password1: string;
+}
