+++ frontend/vite.config.ts
@@ -10,5 +10,8 @@ export default defineConfig({
     watch: {
       usePolling: true
     }
+  },
+  optimizeDeps: {
+    exclude: ['locomotive-scroll']
   }
 })
